Apply safe PR review fixes for venue geocoding.
Load venue-map CSS for address-only pages, move shared geocoding into
base/services/geo.py, and read cache before provider availability checks
without changing working geocoding behavior.
